feat: junção de componentes na tela principal e ajustes nas proporções do footer
