--- Ponte/src/Pages/Footer/footer.jsx
@@ -1,17 +1,0 @@
-import * as S from './footer.style';
-import logo from '../img/LogoPonte5.png';
-
-function Footer() {
-  return (
-    <>
-      <S.Container>
-        <div>
-          <img src={logo} alt="" />
-        </div>
-        
-      </S.Container>
-    </>
-  );
-}
-
-export default Footer;
