+++ Ponte/src/Pages/Home.jsx
@@ -3,6 +3,7 @@ import SobreNos from "./SobreNos/SobreNos";
 import NavBar from "../Components/organisms/NavBar";
 import MainSec from "./LP-index/Sessão inicial";
 import Footer from "../Components/molecules/Footer/Footer";
+import Funcionalidades from "../Pages/LP-index/Funcionalidades/Funcionalidades";
 import Faq from "../Components/organisms/FAQ";
 function Home() {
   return (
@@ -11,6 +12,7 @@ function Home() {
       <NavBar />
       <MainSec />
       <SobreNos />
+      <Funcionalidades />
       <Faq />
       <Footer />
     </>
